// packages/hpl-character-manager/generated/types/CharacterSchema.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as web3 from '@solana/web3.js'
import * as beetSolana from '@metaplex-foundation/beet-solana'
import * as beet from '@metaplex-foundation/beet'
import { CharacterSource, characterSourceBeet } from './CharacterSource'
import { CharacterUsedBy, characterUsedByBeet } from './CharacterUsedBy'
export type CharacterSchema = {
  owner: web3.PublicKey
  source: CharacterSource
  usedBy: CharacterUsedBy
}

/**
 * @category userTypes
 * @category generated
 */
export const characterSchemaBeet =
  new beet.FixableBeetArgsStruct<CharacterSchema>(
    [
      ['owner', beetSolana.publicKey],
      ['source', characterSourceBeet],
      ['usedBy', characterUsedByBeet],
    ],
    'CharacterSchema'
  )